move file types from PD to shared-data
