+++ protocol-designer/src/file-data/types.js
@@ -1,5 +1,5 @@
 // @flow
-import type {ProtocolFile} from '../file-types'
-export type FileMetadataFields = $PropertyType<ProtocolFile<{}>, 'metadata'>
+import type {ProtocolFileV1} from '@opentrons/shared-data'
+export type FileMetadataFields = $PropertyType<ProtocolFileV1<{}>, 'metadata'>
 
 export type FileMetadataFieldAccessors = $Keys<FileMetadataFields>
